requirements.txt
asyncpg==0.30.0
cryptography==44.0.2
fastapi==0.115.12
pydantic==2.11.3
python-dotenv==1.1.0
uvicorn==0.34.0
